server/flask-server/requirements.txt
flask==2.3.3
flask-cors==4.0.0
vosk==0.3.45
spacy==3.7.2
TTS==0.17.6
python-dotenv==1.0.0
gunicorn==21.2.0
pyttsx3==2.90
SpeechRecognition==3.10.0
pydub==0.25.1
gTTS==2.3.2
requests==2.31.0 